refactor: Remove duplication of etcd_ctx in manager.cli
