+++ src/ai/backend/manager/cli/image.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import asyncio
 import logging
